refactor: unify product catalog and deduplicate merch data
- Normalize product data model with `ProductCatalogItem` and `ProductSource`.
- Centralize merch and affiliate gear logic in `src/lib/productCatalog.ts`.
- Extract shared `ProductCard`, `ProductRail`, and `ProductDisclosure` components.
- Refactor Merch, Gear (Toolbox), and Event Guide pages to consume the unified catalog.
- Remove legacy `merch.ts` and duplicate UI components.
- Update SEO schema to derive from catalog data.
