+++ src/data/products/merch.ts
@@ -5,6 +5,13 @@
 
 import { ProductCatalogItem } from './catalog';
 
+export const COLLECTIONS = [
+  { id: 'all', label: 'All' },
+  { id: 'lead-follow-switch', label: 'Lead/Follow/Switch' },
+  { id: 'norcal-bestcal', label: 'NorCal BestCal' },
+  { id: 'rainbow-pride', label: 'Rainbow Pride' },
+] as const;
+
 export const MERCH_PRODUCTS: ProductCatalogItem[] = [
   {
     id: 'love-neon-follow',
