+++ src/pages/Merch.tsx
@@ -5,8 +5,7 @@ import { Box, Stack, Grid, Text, Button } from '@/layouts/Primitives';
 import { SEO } from '@/components/SEO';
 import { ReferralBanner } from '@/components/ReferralBanner';
 import { PageHeader } from '@/components/ui/PageHeader';
-import { COLLECTIONS } from '@/data/merch';
-import { getMerchByCollection, getAllMerchProducts } from '@/lib/productCatalog';
+import { getMerchByCollection, getAllMerchProducts, COLLECTIONS } from '@/lib/productCatalog';
 import { generateMerchSchema } from '@/utils/schema';
 import { cn } from '@/lib/utils';
 import { stroke } from '@/styles/design-tokens';
